# Soft Light: speedup and reduce memory usage