This speeds up and reduces memory usage of soft light tool.

Memory usage is reduced by 12 * width * height byte
Speedup is ~2x

```
12 MP file before: 240 ms, after 120 ms
50 MP file before: 970 ms, after 485 ms
```